requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
google-generativeai==0.7.2
chromadb==0.5.0
pypdf==4.2.0
python-docx==1.1.2
pandas==2.1.4
openpyxl==3.1.4
reportlab==4.2.0
httpx==0.27.0
loguru==0.7.2
tenacity==8.3.0
python-jose==3.3.0
